requirements.txt
# AI Marketing Agent - Dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
